Move swaybg background rendering into common/
swaylock will use it too
